+++ app/src/main/java/com/lunarlog/data/DailyLogFts.kt
@@ -3,6 +3,7 @@ package com.lunarlog.data
 import androidx.room.Entity
 import androidx.room.Fts4
 import androidx.room.ColumnInfo
+import com.lunarlog.core.model.DailyLog
 
 @Entity(tableName = "daily_logs_fts")
 @Fts4(contentEntity = DailyLog::class)